refactor: move signOut.ts into lib/
