+++ app/page.tsx
@@ -1,4 +1,3 @@
-import { signOutAction } from '@/actions/signOut'
 import { auth } from '@/auth'
 import { RoleBadge } from '@/components/general/RoleBadge'
 import AdminControlPanel from '@/components/homepage/AdminControlPanel'
@@ -6,6 +5,7 @@ import SelectAdmissionsCycle from '@/components/homepage/SelectAdmissionsCycle'
 import SignOutButton from '@/components/homepage/SignOutButton'
 import { isSuperUser } from '@/lib/access'
 import { allAllowedAdmissionsCycles, getAllAdminAndTutorEmails } from '@/lib/query/users'
+import { signOutAction } from '@/lib/signOut'
 import { Card, Container, Flex, Heading, Separator } from '@radix-ui/themes'
 import { redirect } from 'next/navigation'
 import React from 'react'
